Revert "f2fs: support 3.10"
This reverts commit b5c96305554df52e71c683228defd6a8215aff59.

+++ fs/f2fs/shrinker.c
@@ -61,7 +61,8 @@ unsigned long f2fs_shrink_count(struct shrinker *shrink,
 	return count;
 }
 
-int f2fs_shrink_scan(struct shrinker *shrink, struct shrink_control *sc)
+unsigned long f2fs_shrink_scan(struct shrinker *shrink,
+				struct shrink_control *sc)
 {
 	unsigned long nr = sc->nr_to_scan;
 	struct f2fs_sb_info *sbi;
